backend/requirements.txt
# Flask & API
flask>=3.0.0
flask-cors>=4.0.0
flask-restx>=1.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
torch>=2.1.0
imbalanced-learn>=0.11.0

# YOLO / Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Explainability
shap>=0.43.0
lime>=0.2.0

# Model Serialization
joblib>=1.3.0
onnx>=1.15.0
onnxruntime>=1.16.0

# Validation & Calibration
calibration-belt>=0.1.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Database (optional)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0
marshmallow>=3.20.0

# Security
python-jose>=3.3.0
cryptography>=41.0.0
